moved schedule_notification_task from tasks to moretasks to fix circular import
