+++ ddpui/celeryworkers/tasks.py
@@ -1,21 +1,19 @@
 """these are tasks which we run through celery"""
 
-import pytz
 import os
 import shutil
 from pathlib import Path
 from datetime import datetime, timedelta
 from subprocess import CompletedProcess
+import pytz
 
 import yaml
 from celery.schedules import crontab
 from django.utils.text import slugify
-from ddpui.auth import ACCOUNT_MANAGER_ROLE, PIPELINE_MANAGER_ROLE
+from ddpui.auth import ACCOUNT_MANAGER_ROLE
 from ddpui.celery import app
-from ddpui.models.notifications import Notification
-from ddpui.models.userpreferences import UserPreferences
 
-from ddpui.utils.sendgrid import send_email_notification, send_schema_changes_email
+
 from ddpui.utils import timezone, awsses
 from ddpui.utils.helpers import find_key_in_dictionary
 from ddpui.utils.custom_logger import CustomLogger
@@ -187,7 +185,7 @@ def setup_dbtworkspace(self, org_id: int, payload: dict) -> str:
     # this client'a dbt setup happens here
     org_dir = DbtProjectManager.get_org_dir(org)
 
-    # four parameters here is correct despite vscode thinking otherwise
+    # five parameters here is correct despite vscode thinking otherwise
     dbtcloned_repo_path = clone_github_repo(
         org.slug,
         payload["gitrepoUrl"],
@@ -1057,20 +1055,3 @@ def setup_periodic_tasks(sender, **kwargs):
         check_org_plan_expiry_notify_people.s(),
         name="check org plan expiry and notify the right people",
     )
-
-
-@app.task(bind=True)
-def schedule_notification_task(self, notification_id, recipient_id):
-    """send scheduled notifications"""
-    notification = Notification.objects.get(id=notification_id)
-    recipient = OrgUser.objects.get(user_id=recipient_id)
-    user_preference, _ = UserPreferences.objects.get_or_create(orguser=recipient)
-
-    notification.sent_time = timezone.as_utc(datetime.now())
-    notification.save()
-
-    if user_preference.enable_email_notifications:
-        try:
-            send_email_notification(user_preference.orguser.user.email, notification.message)
-        except Exception as e:
-            logger.error(f"Error sending discord notification: {str(e)}")
